Report refactoring
